Working Profile Fragment with profile picture implementation.
Fixed issue where app crashed due to Firebase not being able to map User model due to case-sentitive issues.

Changed HomeActivity.java to switched to Official Profile Fragment, and not "ProfileNavFragment" which was our test fragment.

Implemented a dependency to gradle file to allow for Image generation to the ProfileFragment.java (read code for more info).