Fully working Edit profile.
Fixed the image issue by doing the plan explained in the previous commit.

Aditionally, I changed the logic behind this class by making it much simpler. After seeing that I was working with only strings from our DB, I was able to just make a singular updateProfile method that basically took the strings of the user, and changed them accordingly.

Also fixed some code so it looks nicer, and not as comboluted.

Missing comments.